Refactor the basic data in protocol parts out
